# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Rzahan Academy API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Profile {
  id: number;
  userId: string;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  consciousnessLevel?: number | null;
  /** @nullable */
  consciousnessStage?: string | null;
  /** @nullable */
  bio?: string | null;
  createdAt: string;
}

export interface ProfileInput {
  firstName?: string;
  lastName?: string;
  avatarUrl?: string;
  bio?: string;
}

export interface TestResult {
  id: number;
  userId: string;
  totalScore: number;
  stage: number;
  /** @nullable */
  stageName?: string | null;
  /** Per-section scores (JSON) */
  scores?: unknown;
  /** @nullable */
  recommendations?: string | null;
  createdAt: string;
}

export interface TestResultInput {
  /** Array of 40 answers (1-5 scale) */
  answers: number[];
}

export interface Certificate {
  id: number;
  userId: string;
  stage: number;
  /** @nullable */
  stageName?: string | null;
  /** @nullable */
  certificateCode?: string | null;
  issuedAt: string;
}

export interface CertificateInput {
  stage: number;
}

export interface Comment {
  id: number;
  authorName: string;
  /** @nullable */
  avatarUrl?: string | null;
  content: string;
  /** @nullable */
  stage?: number | null;
  /** @nullable */
  stageName?: string | null;
  /** @nullable */
  rating?: number | null;
  createdAt: string;
}

export interface CommentInput {
  content: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating?: number;
}

export interface PlatformStats {
  participants: number;
  completedTests: number;
  issuedCertificates: number;
  activeUsers: number;
}

export interface DashboardStats {
  totalTests: number;
  /** @nullable */
  latestStage?: number | null;
  /** @nullable */
  latestStageName?: string | null;
  progressPercent: number;
  hasCertificate: boolean;
  streakDays: number;
}

export interface JournalEntry {
  id: number;
  userId: string;
  title: string;
  content: string;
  category: string;
  /** @nullable */
  mood?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface JournalEntryInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  content: string;
  category?: string;
  mood?: string;
}

export interface TodayTasks {
  date: string;
  doneSlots: string[];
}

export interface TaskToggleResult {
  done: boolean;
}

export interface LeaderboardEntry {
  userId: string;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  consciousnessLevel?: number | null;
  /** @nullable */
  consciousnessStage?: string | null;
  streak?: number;
  tasksCompleted?: number;
  testCount?: number;
  noteCount?: number;
  badgeCount?: number;
  progressPercent?: number;
  activityScore: number;
}