# lib/api-zod/src/generated/types/comment.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Rzahan Academy API
 * OpenAPI spec version: 0.1.0
 */

export interface Comment {
  id: number;
  authorName: string;
  /** @nullable */
  avatarUrl?: string | null;
  content: string;
  /** @nullable */
  stage?: number | null;
  /** @nullable */
  stageName?: string | null;
  /** @nullable */
  rating?: number | null;
  createdAt: Date;
}