# lib/api-zod/src/generated/types/journalEntry.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Rzahan Academy API
 * OpenAPI spec version: 0.1.0
 */

export interface JournalEntry {
  id: number;
  userId: string;
  title: string;
  content: string;
  category: string;
  /** @nullable */
  mood?: string | null;
  createdAt: Date;
  updatedAt: Date;
}