# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Rzahan Academy API
 * OpenAPI spec version: 0.1.0
 */

export interface Profile {
  id: number;
  userId: string;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  consciousnessLevel?: number | null;
  /** @nullable */
  consciousnessStage?: string | null;
  /** @nullable */
  bio?: string | null;
  createdAt: Date;
}