# lib/api-zod/src/generated/types/todayTasks.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Rzahan Academy API
 * OpenAPI spec version: 0.1.0
 */

export interface TodayTasks {
  date: string;
  doneSlots: string[];
}